# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "ClarificationResponse": {
        "description": "Response body for the owner-scoped clarification read (FTY-152, FTY-170).\n\nCarries a ``needs_clarification`` event's **unanswered** clarification\nquestions ordered by ``position``. The read is status-gated: an owned event\nin any other status \u2014 or one with no unanswered rows \u2014 yields an empty list,\nand the two cases are indistinguishable (no status oracle).",
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestionDTO"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "questions"
        ],
        "title": "ClarificationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogEventCreateRequest": {
        "additionalProperties": false,
        "description": "Request body for ``POST /api/users/{user_id}/log-events``.\n\n``raw_text`` must be non-empty after trimming and at most\n:data:`MAX_RAW_TEXT_LENGTH` characters; surrounding whitespace is stripped so\na whitespace-only entry is rejected rather than stored as blank.\n\n``idempotency_key`` is optional. When present it is an opaque client token,\ntrimmed, non-empty after trimming, and at most\n:data:`MAX_IDEMPOTENCY_KEY_LENGTH` characters; the server stores it verbatim\nand never interprets it. Omitting it preserves the original create behaviour.",
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "raw_text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "raw_text"
        ],
        "title": "LogEventCreateRequest",
        "type": "object"
      },
      "LogEventDTO": {
        "description": "Response body for the log-event create/list/get API.\n\nBuilt directly from the :class:`~app.models.log_events.LogEvent` ORM row.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/LogEventStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "raw_text",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "LogEventDTO",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "fatty-backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/users/{user_id}/log-events": {
      "get": {
        "description": "List the caller's own events for ``day`` (their local timezone).",
        "operationId": "list_log_events_api_users__user_id__log_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "description": "Calendar day (YYYY-MM-DD) in the user's timezone; defaults to today.",
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Calendar day (YYYY-MM-DD) in the user's timezone; defaults to today.",
              "title": "Day"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LogEventDTO"
                  },
                  "title": "Response List Log Events Api Users  User Id  Log Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Log Events",
        "tags": [
          "log-events"
        ]
      },
      "post": {
        "description": "Create a ``pending`` log event and enqueue its estimation job.\n\nThe event is committed first; only then is the job published, so the worker\nnever races ahead of a persisted event. The payload carries ids only \u2014 never\nthe raw text.\n\nWith an ``idempotency_key`` the submit is safe to retry (FTY-096): a fresh\ncreate returns ``201`` and enqueues one job; an idempotent replay returns\n``200`` with the existing event at its current status and enqueues nothing.",
        "operationId": "create_log_event_api_users__user_id__log_events_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogEventCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEventDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Log Event",
        "tags": [
          "log-events"
        ]
      }
    },
    "/api/users/{user_id}/log-events/label": {
      "post": {
        "description": "Create a label event from an uploaded image and extract it in-request (FTY-064).\n\nThe image is the raw request body; its declared type is the ``Content-Type``\nheader. It is validated as **data** (size, content-type allowlist, magic-number\nsignature) at the trust boundary before any work \u2014 an invalid upload fails\nclosed with ``413``/``415`` and no event is created. A valid upload creates a\n``pending`` event, then the FTY-061 label pipeline resolves it synchronously:\nthe raw image only ever lives in this request (it is never enqueued), and is\nretained as a ``log_attachment`` only when ``save`` is set, discarded by\ndefault. Returns the event at its post-extraction status. Errors carry only an\nHTTP status \u2014 never image bytes or extracted content.",
        "operationId": "upload_label_event_api_users__user_id__log_events_label_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "description": "Retain the raw image as a log_attachment (FTY-077). Defaults to off.",
            "in": "query",
            "name": "save",
            "required": false,
            "schema": {
              "default": false,
              "description": "Retain the raw image as a log_attachment (FTY-077). Defaults to off.",
              "title": "Save",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "content-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content-Type"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEventDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Label Event",
        "tags": [
          "log-events"
        ]
      }
    },
    "/api/users/{user_id}/log-events/{event_id}": {
      "get": {
        "description": "Return one of the caller's own events by id, or ``404`` otherwise.",
        "operationId": "get_log_event_api_users__user_id__log_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEventDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Log Event",
        "tags": [
          "log-events"
        ]
      }
    },
    "/api/users/{user_id}/log-events/{event_id}/clarification": {
      "get": {
        "description": "Return the open clarification questions for one of the caller's events.\n\nThe mobile clarify sheet (FTY-153) fetches this lazily when it opens, so the\nToday list/poll DTO stays lean. Questions are ordered by ``position`` and\ncarry their stable ``id`` (the key an answer submission references), their\n``text``, and their quick-pick ``options``. Model-raised parse\nclarifications carry options; deterministic backend-raised questions may\ncarry an empty list, in which case the client shows free-text only.\n\nThe read is status-gated (``log-events.md`` v4): only a\n``needs_clarification`` event serves questions, and only its **unanswered**\nones, so a served question is always freshly answerable. Ownership is\nfail-closed (FTY-030): a cross-user or nonexistent ``event_id`` is\nindistinguishable as ``404`` (no existence oracle); a wrong-status event and\none with no rows both return an empty list (no status oracle). The question\ntext, like raw text, is never logged.",
        "operationId": "get_log_event_clarification_api_users__user_id__log_events__event_id__clarification_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Log Event Clarification",
        "tags": [
          "log-events"
        ]
      }
    },
    "/api/users/{user_id}/log-events/{event_id}/clarification/answers": {
      "post": {
        "description": "Resolve one clarification question on the caller's own event (FTY-171).\n\nThe answer \u2014 a tapped quick-pick option's value or free text \u2014 is applied as\na structured detail to the **same** event, which is re-estimated with every\ndetail answered so far (``needs_clarification \u2192 processing``; the client\npolls as usual). The raw phrase is never mutated and no second event is ever\ncreated (audit findings A3/A5); an empty/whitespace answer is rejected\n``422`` at the request boundary before any work.\n\nA fresh resolve returns ``201`` and enqueues the re-estimate only after the\nanswer, the re-opened job, and the transition are committed \u2014 the worker\nnever races an unpersisted resolve. An idempotent replay of an\nalready-answered question returns ``200`` with the event's current DTO and\nenqueues nothing. A fresh answer for an event that has since moved on\nreturns ``409 {\"error\": \"not_awaiting_clarification\"}`` and mutates nothing.\nOwnership fails closed (``404``, no existence oracle) for a cross-user or\nnonexistent event or a question that is not the event's own. The answer\ntext, like raw text, is never logged.",
        "operationId": "answer_log_event_clarification_api_users__user_id__log_events__event_id__clarification_answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarificationAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEventDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Log Event Clarification",
        "tags": [
          "log-events"
        ]
      }
    },
...
  }
}